+++ FirebaseAppCheck/Sources/Core/Errors/FIRAppCheckHTTPError.m
@@ -17,13 +17,14 @@
 #import "FirebaseAppCheck/Sources/Core/Errors/FIRAppCheckHTTPError.h"
 
 #import "FirebaseAppCheck/Sources/Core/Errors/FIRAppCheckErrorUtil.h"
+#import "FirebaseAppCheck/Sources/Public/FirebaseAppCheck/FIRAppCheckErrors.h"
 
 @implementation FIRAppCheckHTTPError
 
 - (instancetype)initWithHTTPResponse:(NSHTTPURLResponse *)HTTPResponse
                                 data:(nullable NSData *)data {
   NSDictionary *userInfo = [[self class] userInfoWithHTTPResponse:HTTPResponse data:data];
-  self = [super initWithDomain:kFIRAppCheckErrorDomain
+  self = [super initWithDomain:FIRAppCheckErrorDomain
                           code:FIRAppCheckErrorCodeUnknown
                       userInfo:userInfo];
   if (self) {