+++ FirebaseCombineSwift/Sources/Auth/PhoneAuthProvider+Combine.swift
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#if canImport(Combine) && swift(>=5.0) && canImport(FirebaseAuth)
+#if canImport(Combine) && swift(>=5.0)
 
   #if os(iOS)
 
@@ -144,4 +144,4 @@
 
   #endif // os(iOS)
 
-#endif // canImport(Combine) && swift(>=5.0) && canImport(FirebaseAuth)
+#endif // canImport(Combine) && swift(>=5.0)