+++ Firestore/Source/API/FIRFirestore.mm
@@ -198,6 +198,9 @@ NS_ASSUME_NONNULL_BEGIN
   if (!collectionPath) {
     ThrowInvalidArgument("Collection path cannot be nil.");
   }
+  if (!collectionPath.length) {
+    ThrowInvalidArgument("Collection path cannot be empty.");
+  }
   if ([collectionPath containsString:@"//"]) {
     ThrowInvalidArgument("Invalid path (%s). Paths must not contain // in them.", collectionPath);
   }
@@ -210,6 +213,9 @@ NS_ASSUME_NONNULL_BEGIN
   if (!documentPath) {
     ThrowInvalidArgument("Document path cannot be nil.");
   }
+  if (!documentPath.length) {
+    ThrowInvalidArgument("Document path cannot be empty.");
+  }
   if ([documentPath containsString:@"//"]) {
     ThrowInvalidArgument("Invalid path (%s). Paths must not contain // in them.", documentPath);
   }
@@ -221,6 +227,9 @@ NS_ASSUME_NONNULL_BEGIN
 - (FIRQuery *)collectionGroupWithID:(NSString *)collectionID {
   if (!collectionID) {
     ThrowInvalidArgument("Collection ID cannot be nil.");
+  }
+  if (!collectionID.length) {
+    ThrowInvalidArgument("Collection ID cannot be empty.");
   }
   if ([collectionID containsString:@"/"]) {
     ThrowInvalidArgument("Invalid collection ID (%s). Collection IDs must not contain / in them.",
@@ -434,14 +443,13 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (void)getQueryNamed:(NSString *)name completion:(void (^)(FIRQuery *_Nullable query))completion {
   auto firestore = _firestore;
-  auto callback = [completion, firestore](absl::optional<core::Query> query) {
+  auto callback = [completion, firestore](core::Query query, bool found) {
     if (!completion) {
       return;
     }
 
-    if (query.has_value()) {
-      FIRQuery *firQuery = [[FIRQuery alloc] initWithQuery:std::move(query.value())
-                                                 firestore:firestore];
+    if (found) {
+      FIRQuery *firQuery = [[FIRQuery alloc] initWithQuery:std::move(query) firestore:firestore];
       completion(firQuery);
     } else {
       completion(nil);