+++ FirebaseAppCheck/CHANGELOG.md
@@ -1,4 +1,9 @@
-# 8.1.0 -- M97
+# v8.4.0 -- M100
+- [fixed] Bump Promises dependency. (#8365)
+# v8.3.0 -- M99
+- [added] Token API for 3P use. (#8266)
+# v8.2.0 -- M98
 - [added] Apple's App Attest attestation provider support. (#8133)
+- [changed] Token auto-refresh optimizations. (#8232)
 # v8.0.0 -- M95
 - [added] Firebase abuse reduction support SDK. (#7928, #7937, #7948)